Never allow Relax & Autopilot mods
Even if config allows relax & autopilot, in most cases the user does not want people to be able to submit scores with both those mods enabled. If for some reason the user want to allow this then that has to be modified in the code as no config will be added to allow this.